Rename CompressStep to CompressFiles and move it into transport package
- also auto-reformat code fitting to default-style
