+++ xyz-jobs/xyz-job-steps/src/main/java/com/here/xyz/jobs/steps/execution/SyncLambdaStep.java
@@ -22,7 +22,7 @@ package com.here.xyz.jobs.steps.execution;
 import static com.here.xyz.jobs.steps.execution.LambdaBasedStep.ExecutionMode.SYNC;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.here.xyz.jobs.steps.impl.CompressStep;
+import com.here.xyz.jobs.steps.impl.transport.CompressFiles;
 import com.here.xyz.jobs.steps.resources.Load;
 import com.here.xyz.util.service.BaseHttpServerVerticle.ValidationException;
 import java.util.List;
@@ -31,7 +31,7 @@ import java.util.List;
  * A simplified synchronous version of the {@link LambdaBasedStep}.
  */
 @JsonSubTypes({
-        @JsonSubTypes.Type(value = CompressStep.class)
+    @JsonSubTypes.Type(value = CompressFiles.class)
 })
 public abstract class SyncLambdaStep extends LambdaBasedStep<SyncLambdaStep> {
